New datestring avoids error in vim quickfix-window
The Makefile outputs the duration of both compiling and linking, but includes
colons in the output. YCM for vim tries to interpret these as seperators for
compiler output, but can't find the corresponding files. It opens a new buffer
instead of going to the actual location of eventual warnings output by the
compiler. Very weird, hard to explain.