Interactive mode handles arguments recursively.
Previously, inputting `foo bar baz` interactively, the editor would try to
interpret this a single command. It would fail to execute the function
corresponding to `foo` if there was one because `foo` != `foo [...]`.

Now, the editor tries to evaluate `foo` as a single command. It fails if there
is no function that handles command `foo`. Else said function is called and
passed `bar baz`, which again parses all arguments recursively.