Make string-decoding safer.
Sometimes, strings in the save are not terminated but are fixed length. For
example:

The field for a Pokémon's nickname is 10 byte wide. If the nickname is shorter
than 10 characters, the name is followed by a terminating byte (0xff):

    Nickname: "PIDGEY"
          1    2    3    4    5    6    7      8     9    10
        {'P', 'I', 'D', 'G', 'E', 'Y', 0xff, 0x00, 0x00, 0x00} [*]

But if the name spans exactly 10 characters, it has no terminating byte:

    Nickname: "CHARMANDER"
          1    2    3    4    5    6    7    8    9   10
        {'C', 'H', 'A', 'R', 'M', 'A', 'N', 'D', 'E', 'R'}

This makes it unsafe to soley rely on a terminating byte when decoding, thus
making it necessary to set a length after which the decoding should stop (see
commit 569a647c3).

With this in mind, one has to -- when decoding strings from the save --
actually allocate one extra byte in the target C-string just to have space for
a NUL-terminator, should the encoded string not have one. The decoding function
does not place one, as it would have to place it after the bytes it was
requested to decode[**].

    WRONG:
        char charmanders_name[10]; //no room for a NUL-terminator

    RIGHT:
        char charmanders_name[11] = {'\0'};
                                   //room for the whole decoded string,
                                   //plus one byte to place a '\0', which
                                   //conveniently is already placed there
                                   //because of careful pre-initialization.

[*]:  All characters enclosed by single quotes correspond to their
      representation in the proprietary character format, e.g. 'P' = 0xCA.
[**]: It might in the future. God knows how to make this stuff safer. I'll try
      my best.